Changes to calibration

+++ DTO/KalibreringsDTO.cs
@@ -9,7 +9,7 @@ namespace DTO
     public class KalibreringsDTO
     {
         public List<int> ExpectedValue { get; set; }
-        public List<int> ActualValue { get; set; }
+        public List<double> ActualValue { get; set; }
         public DateTime Time { get; set; }
         public string Technician { get; set; }
